Extract CellInfo module and refactor Cell class (#14)
- Move floating methods (info, info_verbose, data) into CellInfo module
- Extract read_note_period, read_effect_command, read_effect_argument
  as private methods in Cell
- Rename data to format_data to avoid collision with Sample#data
- Fix 7 Layout/LineLength offenses (long comments replaced by clean code)
