+++ lib/dr_mod_tracker/dr_mod.rb
@@ -17,6 +17,7 @@ require "lib/dr_mod_tracker/song/song_setup.rb"
 require "lib/dr_mod_tracker/song/song_print.rb"
 require "lib/dr_mod_tracker/song.rb"
 require "lib/dr_mod_tracker/cell/cell_bin.rb"
+require "lib/dr_mod_tracker/cell/cell_info.rb"
 require "lib/dr_mod_tracker/cell.rb"
 require "lib/dr_mod_tracker/pattern/pattern_print.rb"
 require "lib/dr_mod_tracker/pattern.rb"
